zero-config web support
