+++ src/font-family-monospace.ios.tsx
@@ -1,0 +1,1 @@
+export default 'Courier';
