+++ src/font-family-monospace.tsx
@@ -1,0 +1,3 @@
+const FONT_FAMILY_MONOSPACE = 'monospace';
+
+export default FONT_FAMILY_MONOSPACE;
