+++ src/index.tsx
@@ -1,4 +1,5 @@
 export {default as MarkdownTextInput} from './MarkdownTextInput';
-export type {MarkdownTextInputProps, MarkdownStyle} from './MarkdownTextInput';
+export type {MarkdownTextInputProps} from './MarkdownTextInput';
+export type {MarkdownStyle} from './MarkdownTextInput.native'; // .native is fine if we use "export type"
 export type {MarkdownType, MarkdownRange} from './commonTypes';
 export {default as parseExpensiMark} from './parseExpensiMark';
