+++ src/is-web.native.tsx
@@ -1,0 +1,3 @@
+const IS_WEB = false;
+
+export default IS_WEB;
