+++ src/is-web.tsx
@@ -1,0 +1,3 @@
+const IS_WEB = true;
+
+export default IS_WEB;
