+++ src/parseExpensiMark.ts
@@ -1,22 +1,20 @@
 'worklet';
 
-import {Platform} from 'react-native';
 import {ExpensiMark} from 'expensify-common';
 import {unescapeText} from 'expensify-common/dist/utils';
 import {decode} from 'html-entities';
 import type {WorkletFunction} from 'react-native-reanimated/lib/typescript/commonTypes';
 import type {MarkdownType, MarkdownRange} from './commonTypes';
-
-function isWeb() {
-  return Platform.OS === 'web';
-}
+import IS_WEB from './is-web';
 
 function isJest() {
   return !!global.process.env.JEST_WORKER_ID;
 }
 
+const IS_DEV = typeof __DEV__ === 'boolean' ? __DEV__ : process.env.NODE_ENV === 'development';
+
 // eslint-disable-next-line no-underscore-dangle
-if (__DEV__ && !isWeb() && !isJest() && (decode as WorkletFunction).__workletHash === undefined) {
+if (IS_DEV && !IS_WEB && !isJest() && (decode as WorkletFunction).__workletHash === undefined) {
   throw new Error(
     "[react-native-live-markdown] `parseExpensiMark` requires `html-entities` package to be workletized. Please add `'worklet';` directive at the top of `node_modules/html-entities/lib/index.js` using patch-package.",
   );
